Further improvements on making Scala sample projects more idiomatic.
