+++ subprojects/docs/src/samples/scala/crossCompilation/src/main/scala/org/gradle/sample/Person.scala
@@ -1,0 +1,5 @@
+package org.gradle.sample
+
+class Person(val name: String) extends PersonLike {
+  override def getName: String = name
+}
