+++ subprojects/docs/src/samples/scala/crossCompilation/src/main/scala/org/gradle/sample/PersonList.scala
@@ -1,0 +1,5 @@
+package org.gradle.sample
+
+object PersonList {
+  def find(name: String): PersonLike = new JavaPerson(name)
+}
