+++ subprojects/docs/src/samples/scala/customizedLayout/src/scala/org/gradle/sample/Person.scala
@@ -1,0 +1,3 @@
+package org.gradle.sample
+
+class Person(val names: List[String]) extends PersonLike
