+++ subprojects/docs/src/samples/scala/customizedLayout/src/scala/org/gradle/sample/PersonLike.scala
@@ -1,0 +1,5 @@
+package org.gradle.sample
+
+trait PersonLike {
+  def names: List[String]
+}
