--- subprojects/docs/src/samples/scala/customizedLayout/src/scala/org/gradle/sample/api/Person.scala
@@ -1,5 +1,0 @@
-package org.gradle.sample.api
-
-trait Person {
-  def names: List[String]
-}
