--- subprojects/docs/src/samples/scala/customizedLayout/src/scala/org/gradle/sample/impl/PersonImpl.scala
@@ -1,5 +1,0 @@
-package org.gradle.sample.impl
-
-import org.gradle.sample.api.Person
-
-class PersonImpl(val names: List[String]) extends Person
