+++ subprojects/docs/src/samples/scala/force/src/main/scala/org/gradle/sample/PersonLike.scala
@@ -1,8 +1,8 @@
-package org.gradle.sample.api
+package org.gradle.sample
 
 import org.apache.commons.collections.list.GrowthList
 
-trait Person {
+trait PersonLike {
   def names: List[String]
 
   def importedList: GrowthList
