--- subprojects/docs/src/samples/scala/force/src/main/scala/org/gradle/sample/impl/PersonImpl.scala
@@ -1,8 +1,0 @@
-package org.gradle.sample.impl
-
-import org.gradle.sample.api.Person
-import org.apache.commons.collections.list.GrowthList;
-
-class PersonImpl(val names: List[String]) extends Person {
-  override def importedList = new GrowthList
-}
