+++ subprojects/docs/src/samples/scala/force/src/test/scala/org/gradle/sample/PersonSpec.scala
@@ -1,15 +1,14 @@
-package org.gradle.sample.impl
+package org.gradle.sample
 
-import org.gradle.sample.api.Person
 import org.junit.runner.RunWith
 import org.scalatest.FunSpec
 import org.scalatest.junit.JUnitRunner
 
 @RunWith(classOf[JUnitRunner])
-class PersonImplSpec extends FunSpec {
+class PersonSpec extends FunSpec {
   describe("person creation") {
     it("should successfully create a person") {
-      val person: Person = new PersonImpl(List("bob", "smith"))
+      def person: PersonLike = new Person(List("bob", "smith"))
       person
     }
   }
