+++ subprojects/docs/src/samples/scala/mixedJavaAndScala/src/main/scala/org/gradle/sample/JavaPerson.java
@@ -1,0 +1,7 @@
+package org.gradle.sample;
+
+public class JavaPerson extends Person {
+    public JavaPerson(String name) {
+        super(name);
+    }
+}
