--- subprojects/docs/src/samples/scala/mixedJavaAndScala/src/main/scala/org/gradle/sample/impl/JavaPerson.java
@@ -1,7 +1,0 @@
-package org.gradle.sample.impl;
-
-public class JavaPerson extends PersonImpl {
-    public JavaPerson(String name) {
-        super(name);
-    }
-}
