--- subprojects/docs/src/samples/scala/mixedJavaAndScala/src/main/scala/org/gradle/sample/impl/PersonImpl.scala
@@ -1,7 +1,0 @@
-package org.gradle.sample.impl
-
-import org.gradle.sample.Person
-
-class PersonImpl(val name: String) extends Person {
-  override def getName: String = name
-}
