--- subprojects/docs/src/samples/scala/mixedJavaAndScala/src/main/scala/org/gradle/sample/impl/PersonList.scala
@@ -1,7 +1,0 @@
-package org.gradle.sample.impl
-
-import org.gradle.sample.Person
-
-object PersonList {
-  def find(name: String): Person = new JavaPerson(name)
-}
