+++ subprojects/docs/src/samples/scala/quickstart/src/main/scala/org/gradle/sample/Person.scala
@@ -1,0 +1,7 @@
+package org.gradle.sample
+
+import org.apache.commons.collections.list.GrowthList
+
+class Person(val names: List[String]) extends PersonLike {
+  override def importedList = new GrowthList
+}
