+++ subprojects/docs/src/samples/scala/quickstart/src/test/scala/org/gradle/sample/PersonSpec.scala
@@ -1,6 +1,5 @@
-package org.gradle.sample.impl
+package org.gradle.sample
 
-import org.gradle.sample.api.Person
 import org.junit.runner.RunWith
 import org.scalatest.FunSpec
 import org.scalatest.junit.JUnitRunner
@@ -9,7 +8,7 @@ import org.scalatest.junit.JUnitRunner
 class PersonSpec extends FunSpec {
   describe("create a person") {
     it("should succeed for scala") {
-      val person: Person = new PersonImpl(List("bob smith"))
+      val person: PersonLike = new Person(List("bob smith"))
       person
     }
   }
